Food,ActivityCatalog,Recipe

+++ Client/src/app/pages/recipes/recipes.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { RecipeResponseDto, RecipesClient } from '../../shared/models/Nswag generated/NswagGenerated';
 import { AccountService } from '../../account/account.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-recipes',
@@ -13,7 +14,7 @@ export class RecipesComponent implements OnInit {
 
   constructor(
     private recipesClient: RecipesClient,
-    public accountService: AccountService
+    public accountService: AccountService, private router: Router
   ) {}
 
   ngOnInit(): void {
@@ -28,16 +29,36 @@ export class RecipesComponent implements OnInit {
   }
 
   onEdit(recipe: RecipeResponseDto): void {
-    // TODO: navigáció a szerkesztő oldalra vagy modal megnyitása
-    console.log('Edit:', recipe);
+    if (recipe.recipeID) {
+      this.router.navigate(['/recipes/edit', recipe.recipeID]);
+  }
   }
 
   onDelete(recipe: RecipeResponseDto): void {
-   /* if (confirm(`Biztosan törlöd a(z) "${recipe.title}" receptet?`)) {
-      this.recipesClient.delete(recipe.recipeID).subscribe({
-        next: () => this.loadRecipes(),
-        error: err => console.error('Törlés sikertelen:', err)
-      });
-    }*/
+    if (!recipe.recipeID) return;
+  
+    const confirmed = confirm(`You wanna delete the "${recipe.title}" receptet?`);
+    if (!confirmed) return;
+  
+    this.recipesClient.delete(recipe.recipeID).subscribe({
+      next: () => {
+        console.log('Recept törölve');
+        this.recipes = this.recipes.filter(r => r.recipeID !== recipe.recipeID);
+      },
+      error: err => {
+        if (err.error instanceof Blob) {
+          err.error.text().then((text: string) => {
+            try {
+              const parsed = JSON.parse(text);
+              alert(parsed.message || 'Nem törölhető. Lehet, hogy használatban van.');
+            } catch {
+              alert('Törlés sikertelen. (nyers válasz)');
+            }
+          });
+        } else {
+          alert('You use this recipe.');
+        }
+      }
+    });
   }
 }
